refactor(ai): apply adapter pattern to AI module
Apply the same adapter pattern established in the core routecraft package
to all AI module adapters (MCP, LLM, Embedding, Agent). This completes
the adapter pattern refactoring across the entire codebase.

**Breaking Changes:** None - Public API unchanged

**What Changed:**

1. **MCP Adapter (Multi-Interface):**
   - Split McpAdapter → McpSourceAdapter + McpDestinationAdapter
   - Moved to `packages/ai/src/mcp/adapters/mcp/` directory
   - Separate classes: source.ts, destination.ts, shared.ts, types.ts
   - Factory uses structural type guards (arguments.length)
   - Integrated client-adapter.ts and source-adapter.ts into new structure

2. **LLM Adapter (Single-Interface):**
   - Renamed LlmAdapter → LlmDestinationAdapter
   - Moved to `packages/ai/src/llm/destination.ts`
   - Factory returns Destination<unknown, LlmResult> interface type

3. **Embedding Adapter (Single-Interface):**
   - Renamed EmbeddingAdapter → EmbeddingDestinationAdapter
   - Moved to `packages/ai/src/embedding/destination.ts`
   - Factory returns Destination<T, EmbeddingResult> interface type

4. **Agent Adapter (Single-Interface):**
   - Renamed AgentAdapter → AgentDestinationAdapter
   - Moved to `packages/ai/src/agent/destination.ts`
   - Factory returns Destination<unknown, AgentResult> interface type

5. **Updated Imports:**
   - All module index.ts files export new adapter names
   - Main packages/ai/src/index.ts updated with new exports
   - All test files updated to use new class names

**Pattern Consistency:**

All AI adapters now follow the same pattern as core adapters:
- Directory structure with operation-based files (source.ts, destination.ts)
- Class naming: {Concept}{Operation}Adapter
- Factories return interface types (not classes)
- Structural type guards (not runtime value checks)
- Single factory per concept (DX preserved)

**Why This Matters:**

- Consistency: All 11 adapters across both packages follow same structure
- Maintainability: Separate classes with single responsibility
- Reviewability: Changes isolated to specific operation files
- Extensibility: Easy to add new operations
- Type Safety: Structural guards catch errors at build time

**DX Preserved:**
Users still write the same code:
- mcp(serverOptions, schema) for Source
- mcp(clientOptions) for Destination
- llm(options), embedding(options), agent(options) unchanged

**Verification:**
✅ 349 tests passed, 0 skipped
✅ pnpm lint && pnpm format && pnpm typecheck && pnpm build && pnpm test
