+++ packages/ai/src/embedding/index.ts
@@ -1,4 +1,4 @@
-export { EmbeddingAdapter } from "./adapter.ts";
+export { EmbeddingDestinationAdapter } from "./destination.ts";
 export { embedding } from "./embedding.ts";
 export { embeddingPlugin } from "./plugin.ts";
 export { disposeEmbeddingPipelineCache } from "./providers/index.ts";
