+++ packages/ai/src/mcp/adapters/mcp/shared.ts
@@ -1,0 +1,5 @@
+/**
+ * Brand symbol for MCP adapter type checking.
+ * Used internally to identify MCP adapters.
+ */
+export const BRAND_MCP_ADAPTER = Symbol.for("routecraft.mcp.adapter");
