+++ packages/ai/src/mcp/adapters/mcp/types.ts
@@ -1,0 +1,22 @@
+import type { StandardSchemaV1 } from "@standard-schema/spec";
+import type { Exchange } from "@routecraft/routecraft";
+
+/** Message type derived from schema S when present; otherwise unknown. */
+export type McpMessage<S extends StandardSchemaV1 | undefined> =
+  S extends StandardSchemaV1 ? StandardSchemaV1.InferOutput<S> : unknown;
+
+/**
+ * Extracts MCP tool arguments from an exchange. Default implementation uses exchange.body.
+ */
+export type McpArgsExtractor = (
+  exchange: Exchange<unknown>,
+) => Record<string, unknown>;
+
+/**
+ * HTTP client config for a remote MCP server (Streamable HTTP).
+ * Used in mcpPlugin({ clients: { name: config } }).
+ */
+export interface McpClientHttpConfig {
+  transport?: "streamable-http";
+  url: string;
+}
